Rely on the /usr/bin/google-chrome script to run chrome.
This allows to properly set parameters and environment variables. It allows
to get rid of the "no sandbox" warning. To get rid of the "no API keys"
warning, be sure to rebuild chrome with "disable-api-keys-info-bar", or
equivalently, edit /usr/bin/google-chrome so that it sets the following
environment variables:

    export GOOGLE_API_KEY="no"
    export GOOGLE_DEFAULT_CLIENT_ID="no"
    export GOOGLE_DEFAULT_CLIENT_SECRET="no"

This commit also removes a useless "--no-sandbox" for content_shell.